requirements.txt
tkinterdnd2
pystray
Pillow
